requirements.txt
ujson
textdistance
matplotlib
wandb==0.12.21
nltk
datasets
seqeval
lxml
pandas
scipy
scispacy
sentencepiece
rouge_score
torch_optimizer
fuzzywuzzy
python-Levenshtein
evaluate
bert-score
accelerate
stemming
quantulum3
deepspeed
openai
dill
scispacy
ujson
p-tqdm
tqdm
scikit-learn==1.0.2
swifter
pytorch-lightning
stanza
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.1/en_core_sci_sm-0.5.1.tar.gz
